Convert college details styles to plain CSS
This is done so that the college details
appear same in Drawer component in small screens.

This is also due to AntD not recognizing SCSS files
inside component where Drawer is used.

The choice is either use LESS or CSS.
To convert to LESS at this stage of development will be a big headache.